Conflicts:
	scripts/msct_parser.py
	scripts/sct_segment_graymatter.py

Former-commit-id: cd2c1a4a9651f7cabe6078a1d9559325e057f73b [formerly d45feb621838cb4033e655b8fd21fc4ccf28c41d] [formerly 0150adf7940552b95e09bf8eb18b91f45c922018 [formerly cbf59e237eb0f8d5e7e0b96e50397db91169147c [formerly 56400a35670cb5ddcc7dddf6a531459b4847715d]]]
Former-commit-id: cd2c1a4a9651f7cabe6078a1d9559325e057f73b [formerly d45feb621838cb4033e655b8fd21fc4ccf28c41d]
Former-commit-id: d39883c7bfa1b86b4d84fbbed253a20924220eed [formerly 758a35fb09bbe0c85969758900d9886e39a4dec8 [formerly 9593aa3b9b0734da189979b51b9ba926c39d5081]]
Former-commit-id: 925027ac82775564198d4c3311359a7eff448b7f
Former-commit-id: ffb47dd27c003c1dad79ab848c1868afe2654a61

+++ dev/template_creation/normalize.sh
@@ -2,7 +2,7 @@
 
 path=$(pwd)
 
-cd /home/django/jtouati/data/template_data/new_data
+cd /home/django/jtouati/data/template_data/
 
 ########################## T2
 
@@ -12,11 +12,11 @@ fi
 
 mkdir t2_output
 
-SUBJECTS_MTL="errsm_02 errsm_04 errsm_26 errsm_31 errsm_32 errsm_33"
+SUBJECTS_MTL="errsm_02 errsm_04 errsm_05 errsm_09 errsm_10 errsm_11 errsm_12 errsm_13 errsm_14 errsm_16 errsm_17 errsm_18 errsm_20 errsm_21 errsm_22 errsm_23 errsm_24 errsm_25 errsm_26 errsm_30 errsm_31 errsm_32 errsm_33"
 
-SUBJECTS_MAR="ALT JD JW MT T047 VC VG VP"
+SUBJECTS_MAR="1 7 ALT JD JW MD MLL MT T047 VC VG VP"
 
-cd Montreal/
+cd montreal/
 
 for i in $SUBJECTS_MTL; do
 
@@ -28,13 +28,21 @@ for i in $SUBJECTS_MTL; do
 	fi
 	mkdir second_step
 	echo "Copying files"
-	cp ${i}_t2_crop_straight_crop_2temp.nii.gz second_step/${i}_preprocessed.nii.gz
+	
+	if [ -e "${i}_t2_crop_straight_crop_2temp.nii.gz" ];then
+		cp ${i}_t2_crop_straight_crop_2temp.nii.gz second_step/${i}_preprocessed.nii.gz
+	fi
+	
+	if [ -e "native2temp.nii.gz" ];then
+		cp native2temp.nii.gz second_step/${i}_preprocessed.nii.gz
+	fi
+	
 	cp centerline_straight_crop_2temp.nii.gz second_step/centerline_straight_${i}.nii.gz
 	cd second_step/
 
 
 	echo "Align_vertebrae"
-	sct_align_vertebrae.py -i ${i}_preprocessed.nii.gz -l ../${i}_labels.nii.gz -R /home/django/jtouati/code/spinalcordtoolbox/dev/template_creation/template_shape-mask.nii.gz -o ${i}_aligned.nii.gz -t affine - w spline
+	sct_align_vertebrae.py -i ${i}_preprocessed.nii.gz -l ../${i}_labels.nii.gz -R /home/django/jtouati/code/spinalcordtoolbox/dev/template_creation/template_shape-mask.nii.gz -o ${i}_aligned.nii.gz -t affine -w spline
 
 	echo "applying transfo to the centerline"
 	WarpImageMultiTransform 3 centerline_straight_${i}.nii.gz centerline_straight_${i}_aligned.nii.gz -R /home/django/jtouati/code/spinalcordtoolbox/dev/template_creation/template_shape-mask.nii.gz n2t.txt --use-BSpline
@@ -46,11 +54,15 @@ for i in $SUBJECTS_MTL; do
 	echo "moving file"
 	mv ${i}_aligned_normalized.nii.gz ../../../../t2_output/${i}_preprocessed.nii.gz
 
-	cd ../../..
+    cd ..
+
+    rm -rf second_step
+
+	cd ../..
 
 done
 
-cd ../Marseille/
+cd ../marseille/
 
 for i in $SUBJECTS_MAR; do
 
@@ -62,13 +74,21 @@ for i in $SUBJECTS_MAR; do
 	fi
 	mkdir second_step
 	echo "Copying files"
-	cp ${i}_t2_crop_straight_crop_2temp.nii.gz second_step/${i}_preprocessed.nii.gz
+	
+	if [ -e "${i}_t2_crop_straight_crop_2temp.nii.gz" ];then
+		cp ${i}_t2_crop_straight_crop_2temp.nii.gz second_step/${i}_preprocessed.nii.gz
+	fi
+	
+	if [ -e "native2temp.nii.gz" ];then
+		cp native2temp.nii.gz second_step/${i}_preprocessed.nii.gz
+	fi
+	
 	cp centerline_straight_crop_2temp.nii.gz second_step/centerline_straight_${i}.nii.gz
 	cd second_step/
 
 
 	echo "Align_vertebrae"
-	sct_align_vertebrae.py -i ${i}_preprocessed.nii.gz -l ../${i}_labels.nii.gz -R /home/django/jtouati/code/spinalcordtoolbox/dev/template_creation/template_shape-mask.nii.gz -o ${i}_aligned.nii.gz -t affine - w spline
+	sct_align_vertebrae.py -i ${i}_preprocessed.nii.gz -l ../${i}_labels.nii.gz -R /home/django/jtouati/code/spinalcordtoolbox/dev/template_creation/template_shape-mask.nii.gz -o ${i}_aligned.nii.gz -t affine -w spline
 
 	echo "applying transfo to the centerline"
 	WarpImageMultiTransform 3 centerline_straight_${i}.nii.gz centerline_straight_${i}_aligned.nii.gz -R /home/django/jtouati/code/spinalcordtoolbox/dev/template_creation/template_shape-mask.nii.gz n2t.txt --use-BSpline
@@ -79,92 +99,97 @@ for i in $SUBJECTS_MAR; do
 	echo "moving file"
 	mv ${i}_aligned_normalized.nii.gz ../../../../t2_output/${i}_preprocessed.nii.gz
 
-	cd ../../..
+    cd ..
+
+    rm -rf second_step
+
+
+	cd ../..
 
 
 done
 
-cd /home/django/jtouati/data/template_data/new_data
-
-
-if [ -e "t1_output" ];then
-    rm -rf t1_output
-fi
-
-mkdir t1_output
-
-SUBJECTS_MTL="errsm_02 errsm_04 errsm_31 errsm_32 errsm_33"
-
-SUBJECTS_MAR="ALT JD JW MT T047 T020b VG VP"
-
-cd Montreal/
-
-for i in $SUBJECTS_MTL; do
-
-	cd $i/T1
-	echo $i
-
-	if [ -e "second_step" ];then
-	    rm -rf second_step
-	fi
-	mkdir second_step
-	echo "Copying files"
-	cp ${i}_t1_crop_straight_crop_2temp.nii.gz second_step/${i}_preprocessed.nii.gz
-	cp centerline_straight_crop_2temp.nii.gz second_step/centerline_straight_${i}.nii.gz
-	cd second_step/
-
-
-	echo "Align_vertebrae"
-	sct_align_vertebrae.py -i ${i}_preprocessed.nii.gz -l ../../T2/${i}_labels.nii.gz -R /home/django/jtouati/code/spinalcordtoolbox/dev/template_creation/template_shape-mask.nii.gz -o ${i}_aligned.nii.gz -t affine - w spline
-
-	echo "applying transfo to the centerline"
-	WarpImageMultiTransform 3 centerline_straight_${i}.nii.gz centerline_straight_${i}_aligned.nii.gz -R /home/django/jtouati/code/spinalcordtoolbox/dev/template_creation/template_shape-mask.nii.gz n2t.txt --use-BSpline
-
-
-	echo "normalizing"
-	sct_normalize.py -i ${i}_aligned.nii.gz -c centerline_straight_${i}_aligned.nii.gz
-
-	echo "moving file"
-	mv ${i}_aligned_normalized.nii.gz ../../../../t1_output/${i}_preprocessed.nii.gz
-
-	cd ../../..
-
-done
-
-cd ../Marseille/
-
-for i in $SUBJECTS_MAR; do
-
-	cd $i/T1
-	echo $i
-
-	if [ -e "second_step" ];then
-	    rm -rf second_step
-	fi
-	mkdir second_step
-	echo "Copying files"
-	cp ${i}_t1_crop_straight_crop_2temp.nii.gz second_step/${i}_preprocessed.nii.gz
-	cp centerline_straight_crop_2temp.nii.gz second_step/centerline_straight_${i}.nii.gz
-	cd second_step/
-
-
-	echo "Align_vertebrae"
-	sct_align_vertebrae.py -i ${i}_preprocessed.nii.gz -l  ../../T2/${i}_labels.nii.gz -R /home/django/jtouati/code/spinalcordtoolbox/dev/template_creation/template_shape-mask.nii.gz -o ${i}_aligned.nii.gz -t affine - w spline
-
-	echo "applying transfo to the centerline"
-	WarpImageMultiTransform 3 centerline_straight_${i}.nii.gz centerline_straight_${i}_aligned.nii.gz -R /home/django/jtouati/code/spinalcordtoolbox/dev/template_creation/template_shape-mask.nii.gz n2t.txt --use-BSpline
-
-	echo "normalizing"
-	sct_normalize.py -i ${i}_aligned.nii.gz -c centerline_straight_${i}_aligned.nii.gz
-
-	echo "moving file"
-	mv ${i}_aligned_normalized.nii.gz ../../../../t1_output/${i}_preprocessed.nii.gz
-
-	cd ../../..
-
-
-done
-
-
+# cd /home/django/jtouati/data/template_data
+#
+#
+# if [ -e "t1_output" ];then
+#     rm -rf t1_output
+# fi
+#
+# mkdir t1_output
+#
+# SUBJECTS_MTL="errsm_02 errsm_04 errsm_31 errsm_32 errsm_33"
+#
+# SUBJECTS_MAR="ALT JD JW MT T047 T020b VG VP"
+#
+# cd Montreal/
+#
+# for i in $SUBJECTS_MTL; do
+#
+# 	cd $i/T1
+# 	echo $i
+#
+# 	if [ -e "second_step" ];then
+# 	    rm -rf second_step
+# 	fi
+# 	mkdir second_step
+# 	echo "Copying files"
+# 	cp ${i}_t1_crop_straight_crop_2temp.nii.gz second_step/${i}_preprocessed.nii.gz
+# 	cp centerline_straight_crop_2temp.nii.gz second_step/centerline_straight_${i}.nii.gz
+# 	cd second_step/
+#
+#
+# 	echo "Align_vertebrae"
+# 	sct_align_vertebrae.py -i ${i}_preprocessed.nii.gz -l ../../T2/${i}_labels.nii.gz -R /home/django/jtouati/code/spinalcordtoolbox/dev/template_creation/template_shape-mask.nii.gz -o ${i}_aligned.nii.gz -t affine - w spline
+#
+# 	echo "applying transfo to the centerline"
+# 	WarpImageMultiTransform 3 centerline_straight_${i}.nii.gz centerline_straight_${i}_aligned.nii.gz -R /home/django/jtouati/code/spinalcordtoolbox/dev/template_creation/template_shape-mask.nii.gz n2t.txt --use-BSpline
+#
+#
+# 	echo "normalizing"
+# 	sct_normalize.py -i ${i}_aligned.nii.gz -c centerline_straight_${i}_aligned.nii.gz
+#
+# 	echo "moving file"
+# 	mv ${i}_aligned_normalized.nii.gz ../../../../t1_output/${i}_preprocessed.nii.gz
+#
+# 	cd ../../..
+#
+# done
+#
+# cd ../Marseille/
+#
+# for i in $SUBJECTS_MAR; do
+#
+# 	cd $i/T1
+# 	echo $i
+#
+# 	if [ -e "second_step" ];then
+# 	    rm -rf second_step
+# 	fi
+# 	mkdir second_step
+# 	echo "Copying files"
+# 	cp ${i}_t1_crop_straight_crop_2temp.nii.gz second_step/${i}_preprocessed.nii.gz
+# 	cp centerline_straight_crop_2temp.nii.gz second_step/centerline_straight_${i}.nii.gz
+# 	cd second_step/
+#
+#
+# 	echo "Align_vertebrae"
+# 	sct_align_vertebrae.py -i ${i}_preprocessed.nii.gz -l  ../../T2/${i}_labels.nii.gz -R /home/django/jtouati/code/spinalcordtoolbox/dev/template_creation/template_shape-mask.nii.gz -o ${i}_aligned.nii.gz -t affine - w spline
+#
+# 	echo "applying transfo to the centerline"
+# 	WarpImageMultiTransform 3 centerline_straight_${i}.nii.gz centerline_straight_${i}_aligned.nii.gz -R /home/django/jtouati/code/spinalcordtoolbox/dev/template_creation/template_shape-mask.nii.gz n2t.txt --use-BSpline
+#
+# 	echo "normalizing"
+# 	sct_normalize.py -i ${i}_aligned.nii.gz -c centerline_straight_${i}_aligned.nii.gz
+#
+# 	echo "moving file"
+# 	mv ${i}_aligned_normalized.nii.gz ../../../../t1_output/${i}_preprocessed.nii.gz
+#
+# 	cd ../../..
+#
+#
+# done
+#
+#
